+++ R/mod_news.R
@@ -17,24 +17,33 @@ mod_news_ui <- function(id){
                           h4(strong("Historique des versions")),
                           br(),
 
-                          bslib::accordion(open = HTML("<strong>Avril 2024</strong>"), # Open only the latest update (item 2)
+                          bslib::accordion(open = c("tocome", "july24"), # refers to 'value' of accordion_panel. update when needed
                                            multiple = TRUE,
 
                                            # Next updates ----
                                            bslib::accordion_panel(title = HTML("<strong>À venir</strong>"),
+                                                                  value = "tocome",
                                                                   icon = bsicons::bs_icon("question-circle", size = "1.5rem"),
 
                                                                   HTML(bsicons::bs_icon("database-fill-add", size = "1.5rem"),
                                                                        "Ajout des données de distribution de gaz 2018-2022"),
                                                                   br(),br(),
                                                                   HTML(bsicons::bs_icon("arrow-clockwise", size = "1.5rem"),
-                                                                       "Mise à jour des données de production d'électricité 2023"),
-                                                                  br(),br(),
-                                                                  HTML(bsicons::bs_icon("arrow-clockwise", size = "1.5rem"),
                                                                        "Mise à jour des données de distribution d'électricité 2022")
                                            ),
+
+                                           # Juillet 2024 ----
+
+                                           bslib::accordion_panel(title = HTML("<strong>Juillet  2024</strong>"),
+                                                                  value = "july24",
+                                                                  icon = bsicons::bs_icon("calendar-check", size = "1.5rem"),
+                                                                  HTML(bsicons::bs_icon("arrow-clockwise", size = "1.5rem"),
+                                                                       "Mise à jour des données de production d'électricité 2023 et révision des années 2015-2022")
+                                           ),
+
                                            # April 2024 ----
                                            bslib::accordion_panel(title = HTML("<strong>Avril 2024</strong>"),
+                                                                  value = "april24",
                                                                   icon = bsicons::bs_icon("calendar-check", size = "1.5rem"),
 
                                                                   HTML(bsicons::bs_icon("database-fill-add", size = "1.5rem"),
@@ -55,6 +64,7 @@ mod_news_ui <- function(id){
 
                                            # June 2023 ----
                                            bslib::accordion_panel(title = HTML("<strong>Juin 2023</strong>"),
+                                                                  value = "june23",
                                                                   icon = bsicons::bs_icon("calendar-check", size = "1.5rem"),
 
                                                                   HTML(bsicons::bs_icon("database-fill-add", size = "1.5rem"),